+++ visualizer/src/visuals/neon_calligraphy.rs
@@ -21,6 +21,7 @@ const MAT_CONFIG: effects::HueMaterialConfig = effects::HueMaterialConfig {
     saturation: 0.9,
     lightness: 0.5,
     emissive_strength: EMISSIVE_STRENGTH,
+    frequency_mapped: false,
 };
 
 #[derive(Component)]